Log-In Done close  #3
