+++ curriculum/src/app/app.module.ts
@@ -14,7 +14,7 @@ import {CollapseModule} from 'ngx-bootstrap';
 import {HeaderComponent} from './core/header/header.component';
 import {PageNotFoundComponent} from './core/page-not-found/page-not-found.component';
 import { CreateAccountComponent } from './core/create-account/create-account.component';
-import { LogInAccountComponent } from './log-in-account/log-in-account.component';
+import { LogInAccountComponent } from './core/log-in-account/log-in-account.component';
 
 const appRoutes: Routes = [
     { path: '',
