+++ curriculum/src/app/core/create-account/create-account.component.html
@@ -14,23 +14,26 @@
                 <div class="form-group row">
                   <label for="inputName" class="col">Name</label>
                   <div class="col">
-                    <input type="text" class="form-control" id="textId" >
+                    <input type="text" class="form-control" id="textId" required>
                   </div>
                 </div>
                 <div class="form-group row">
                   <label for="staticEmail" class="col">Email</label>
                   <div class="col">
-                    <input type="email" class="form-control" id="staticEmail" placeholder="email@example.com">
+                    <input type="email" class="form-control" id="staticEmail" placeholder="email@example.com" required>
                   </div>
                 </div>
                 <div class="form-group row">
                   <label for="inputPassword" class="col">Password</label>
                   <div class="col">
-                    <input type="password" class="form-control" id="inputPassword" placeholder="Password">
+                    <input type="password" class="form-control" id="inputPassword" placeholder="Password" required>
                   </div>
                 </div>
+                <button class="btn btn-primary " type="submit" id="buttonCreate">
+                  <i class="pi pi-pencil"></i>
+                  Create
+                </button>
               </form>
-              <a href="#" class="btn btn-primary" id="buttonCreate">Create</a>
             </li>
           </ul>
         </div>
