+++ curriculum/src/app/core/header/header.component.scss
@@ -23,6 +23,11 @@
   background: lighten(orange, 20%);
   border-color: orange;
 }
+#registerButton:hover{
+  background: lighten(orange, 20%);
+  border-color: orange;
+}
+
 #registerButton{
   background: orange;
   border-color: orange;
