+++ curriculum/src/app/core/log-in-account/log-in-account.component.html
@@ -1,0 +1,64 @@
+<div class="container">
+  <div class="row">
+    <div class="col-sm">
+      <div class="card" id="cardCreate">
+        <div class="card-body">
+          <ul class="list-group list-group-flush">
+            <li class="list-group-item">
+              <div class="radioPrime">
+                <p-radioButton name="groupname" value="val2" (onClick)="redirect('create')"></p-radioButton>
+              </div>
+              <label for="radio"> Create new Account</label>
+            </li>
+            <li class="list-group-item">
+              <h2 class="card-title">First Time here?</h2>
+              <p>
+                Lets create a new account for your. Its totally free.
+              </p>
+              <p>
+                Join Us in this adventure !
+              </p>
+              <button class="btn btn-info " type="button" id="buttonCreate" (click)="redirect('create')">
+                <i class="pi pi-pencil"></i>
+                Create
+              </button>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+    <div class="col-sm">
+      <div class="card" id="cardLogin">
+        <div class="card-body">
+          <ul class="list-group list-group-flush">
+            <li class="list-group-item">
+              <mat-radio-button value="1" [checked]="true" [color]="checkboxColor"></mat-radio-button>
+              <label for="radio"> Log-In to your current Account</label>
+            </li>
+            <li class="list-group-item">
+              <h2 class="card-title">Welcome Back!</h2>
+              <form>
+                <div class="form-group row">
+                  <label for="staticEmail" class="col">Email</label>
+                  <div class="col">
+                    <input type="email" class="form-control" id="staticEmail" placeholder="email@example.com" required>
+                  </div>
+                </div>
+                <div class="form-group row">
+                  <label for="inputPassword" class="col">Password</label>
+                  <div class="col">
+                    <input type="password" class="form-control" id="inputPassword" placeholder="Password" required>
+                  </div>
+                </div>
+                <button class="btn btn-primary " type="submit" id="buttonCreate">
+                  <i class="pi pi-user"></i>
+                  Log-In
+                </button>
+              </form>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
