--- curriculum/src/app/log-in-account/log-in-account.component.html
@@ -1,1 +1,0 @@
-<p>log-in-account works!</p>
