--- curriculum/src/app/log-in-account/log-in-account.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'cod-log-in-account',
-  templateUrl: './log-in-account.component.html',
-  styleUrls: ['./log-in-account.component.scss']
-})
-export class LogInAccountComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
